sep24 commit

+++ src/com/vijay/FactorialNumber.java
@@ -3,13 +3,13 @@ package com.vijay;
 import java.util.Scanner;
 
 public class FactorialNumber {
-//factorial of a number is product of all positive descending integers
+    //factorial of a number is product of all positive descending integers
 	public static void main(String[] args) {
 		Scanner sc = new Scanner(System.in);
 		System.out.print("enter number: ");
 		int n = sc.nextInt();
 
-		int factorial = 1;
+		long factorial = 1;
 		for (int i = 1; i <= n; i++) {
 			factorial = factorial * i;
 		}
